requirements.txt
azure-mgmt-resource==23.0.1
azure-mgmt-web==7.2.0
azure-mgmt-compute==30.4.0
azure-identity==1.5.0
tabulate==0.9.0
click